+++ Wrappers/Python/ccpi/utilities/imagequality.py
@@ -1,0 +1,29 @@
+# -*- coding: utf-8 -*-
+# Copyright 2019 Science Technology Facilities Council
+# Copyright 2019 University of Manchester
+#
+# This work is part of the Core Imaging Library developed by Science Technology
+# Facilities Council and University of Manchester
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0.txt
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from __future__ import print_function, division
+import numpy
+
+def psnr(img1, img2, data_range=1):
+    mse = numpy.mean( (img1 - img2) ** 2 )
+    if mse == 0:
+        return 1000
+    return 20 * numpy.log10(data_range / numpy.sqrt(mse))
+
+
